api/requirements.txt
flask
flask_restful
pika
flask_cors
google-cloud-storage